[EventDispatcher] feat!: use event as first argument
The event dispatcher "dispatch" method has been modified to accept the
event as first (and only required) argument. Additionally the event
object doea not need to extend the jymfony base event class anymore.
The event name can be provided as second argument. If not provided the
class fully qualified name will be used as event name.